+++ tests/core/test_hooks.rs
@@ -1,4 +1,8 @@
+use branchless::core::eventlog::testing::{get_event_replayer_events, redact_event_timestamp};
+use branchless::core::eventlog::{Event, EventLogDb, EventReplayer};
+use branchless::core::formatting::Glyphs;
 use branchless::testing::make_git;
+use branchless::tui::Effects;
 use branchless::util::get_sh;
 use eyre::{eyre, Context};
 use std::process::Command;
@@ -201,6 +205,125 @@ fn test_pre_auto_gc() -> eyre::Result<()> {
     );
     insta::assert_snapshot!(stdout, @"branchless: collecting garbage
 ");
+
+    Ok(())
+}
+
+#[test]
+fn test_merge_commit_recorded() -> eyre::Result<()> {
+    let git = make_git()?;
+
+    if !git.supports_reference_transactions()? {
+        return Ok(());
+    }
+
+    git.init_repo()?;
+    let test1_oid = git.commit_file("test1", 1)?;
+    git.run(&["checkout", "HEAD^"])?;
+    git.commit_file("test2", 2)?;
+    git.run(&["merge", &test1_oid.to_string()])?;
+
+    let effects = Effects::new_suppress_for_test(Glyphs::text());
+    let repo = git.get_repo()?;
+    let conn = repo.get_db_conn()?;
+    let event_log_db = EventLogDb::new(&conn)?;
+    let event_replayer = EventReplayer::from_event_log_db(&effects, &repo, &event_log_db)?;
+    let events: Vec<Event> = get_event_replayer_events(&event_replayer)
+        .into_iter()
+        .cloned()
+        .map(redact_event_timestamp)
+        .collect();
+    insta::assert_debug_snapshot!(events, @r###"
+    [
+        RefUpdateEvent {
+            timestamp: 0.0,
+            event_tx_id: EventTransactionId(
+                1,
+            ),
+            ref_name: "HEAD",
+            old_oid: f777ecc9b0db5ed372b2615695191a8a17f79f24,
+            new_oid: 62fc20d2a290daea0d52bdc2ed2ad4be6491010e,
+            message: None,
+        },
+        RefUpdateEvent {
+            timestamp: 0.0,
+            event_tx_id: EventTransactionId(
+                1,
+            ),
+            ref_name: "refs/heads/master",
+            old_oid: f777ecc9b0db5ed372b2615695191a8a17f79f24,
+            new_oid: 62fc20d2a290daea0d52bdc2ed2ad4be6491010e,
+            message: None,
+        },
+        CommitEvent {
+            timestamp: 0.0,
+            event_tx_id: EventTransactionId(
+                2,
+            ),
+            commit_oid: NonZeroOid {
+                inner: 62fc20d2a290daea0d52bdc2ed2ad4be6491010e,
+            },
+        },
+        RefUpdateEvent {
+            timestamp: 0.0,
+            event_tx_id: EventTransactionId(
+                3,
+            ),
+            ref_name: "HEAD",
+            old_oid: 0000000000000000000000000000000000000000,
+            new_oid: f777ecc9b0db5ed372b2615695191a8a17f79f24,
+            message: None,
+        },
+        RefUpdateEvent {
+            timestamp: 0.0,
+            event_tx_id: EventTransactionId(
+                4,
+            ),
+            ref_name: "HEAD",
+            old_oid: 62fc20d2a290daea0d52bdc2ed2ad4be6491010e,
+            new_oid: f777ecc9b0db5ed372b2615695191a8a17f79f24,
+            message: None,
+        },
+        RefUpdateEvent {
+            timestamp: 0.0,
+            event_tx_id: EventTransactionId(
+                5,
+            ),
+            ref_name: "HEAD",
+            old_oid: f777ecc9b0db5ed372b2615695191a8a17f79f24,
+            new_oid: fe65c1fe15584744e649b2c79d4cf9b0d878f92e,
+            message: None,
+        },
+        CommitEvent {
+            timestamp: 0.0,
+            event_tx_id: EventTransactionId(
+                6,
+            ),
+            commit_oid: NonZeroOid {
+                inner: fe65c1fe15584744e649b2c79d4cf9b0d878f92e,
+            },
+        },
+        RefUpdateEvent {
+            timestamp: 0.0,
+            event_tx_id: EventTransactionId(
+                8,
+            ),
+            ref_name: "HEAD",
+            old_oid: fe65c1fe15584744e649b2c79d4cf9b0d878f92e,
+            new_oid: 91a5ccb4feefba38b0ffa4911c5c3f6c225f662e,
+            message: None,
+        },
+        CommitEvent {
+            timestamp: 0.0,
+            event_tx_id: EventTransactionId(
+                9,
+            ),
+            commit_oid: NonZeroOid {
+                inner: 91a5ccb4feefba38b0ffa4911c5c3f6c225f662e,
+            },
+        },
+    ]
+    "###);
 
     Ok(())
 }