+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed: `git restack` warns if a sub-command fails (e.g. if `git rebase` fails with merge conflicts that need to be resolved).
 - Fixed (#57): `git undo` shows an informative link when dealing with empty events, rather than warning about a bug. Thanks to @waych for reporting.
 - Fixed: Flickering in `git undo`'s rendering has been reduced.
+- Fixed: Commits made via `git merge` are now recorded in the event log.
 
 ## [0.3.4] - 2021-08-12